requirements.txt
# Core Framework
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9

# AI/ML Components
langchain==0.1.9
langchain-community==0.0.24
sentence-transformers==2.3.1
ollama==0.1.6
torch==2.1.0

# Vector Database
qdrant-client==1.7.3

# Document Processing
unstructured==0.12.4
pymupdf==1.23.21
python-docx==1.1.0
markdown==3.5.2

# Text Processing
tiktoken==0.5.2

# Testing Framework
selenium==4.17.2
webdriver-manager==4.0.1

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
numpy==1.24.3